backend/telegram/requirements/requirements.txt
pyTelegramBotApi==4.18.1
PyYAML==6.0.1
pika==1.3.2